style: 카드 호버시 테두리 스타일 변경
