+++ src/components/CardList/CardList.css
@@ -14,6 +14,10 @@
   border-radius: 1.5rem 1.5rem;
 }
 
+.Card:hover {
+  border: 2px solid var(--linkbrary-primary-color, #6d6afe);
+}
+
 .Card-thumbnail {
   width: 100%;
   flex-shrink: 0;
